fix linux support + add gen 3 mode for positional tracking

+++ ZEDCamera/Assets/Samples~/OpenCV ArUco Detection/Scripts/Core/ZEDToOpenCVRetriever.cs
@@ -203,7 +203,7 @@ public class ZEDToOpenCVRetriever : MonoBehaviour
         if(zedmat == null)
         {
             zedmat = new ZEDMat();
-            zedmat.Create(new sl.Resolution((uint)zedCam.ImageWidth, (uint)zedCam.ImageHeight), mattype);
+            zedmat.Create(new sl.Resolution(zedCam.ImageWidth, zedCam.ImageHeight), mattype);
         }
